fix: split factory-graph projection z-axis hint tests for lint limit
Move z-axis incomplete hint assertions into factory-graph-react-flow-projection.z-axis-hints.test.ts so the main projection test file stays under noExcessiveLinesPerFile (450).
