+++ ui/src/features/factory-graph-editor/lib/factory-graph-react-flow-projection.test.ts
@@ -10,7 +10,6 @@ import type {
   CanonicalFactoryDefinition,
   FactoryGraphTopology,
 } from "./factory-graph-draft-types";
-import { getFactoryGraphEditorMessages } from "../messages/editor";
 import { projectFactoryGraphToReactFlow } from "./factory-graph-react-flow-projection";
 
 // biome-ignore lint/complexity/noExcessiveLinesPerFunction: projection contract scenarios stay together around one adapter.
@@ -354,58 +353,6 @@ describe("factory graph React Flow projection", () => {
     );
     expect(anchorIds).not.toContain("workstation-on-continue-source");
     expect(anchorIds).not.toContain("workstation-on-rejection-source");
-    expect(
-      projection.nodes.find((node) => node.id === "workstation:draft")?.data
-        .zAxisIncompleteHints,
-    ).toBeNull();
-  });
-
-  it("sets z-axis incomplete hints when connection editing is enabled without stopWords", () => {
-    const factoryWithoutStopWords = {
-      ...baseFactoryDefinition,
-      workstations: [
-        {
-          ...baseFactoryDefinition.workstations[0],
-          behavior: "STANDARD",
-          stopWords: undefined,
-        },
-      ],
-    } satisfies CanonicalFactoryDefinition;
-    const topology = buildFactoryGraphTopologyFromDefinition(
-      factoryWithoutStopWords,
-    );
-    const hintMessage =
-      getFactoryGraphEditorMessages().zAxisIncompleteConnectionHint;
-
-    const projection = projectFactoryGraphToReactFlow({
-      editor: {
-        canEditConnections: true,
-        pendingAdditionEdgeIds: new Set(),
-        pendingAdditionNodeIds: new Set(),
-        pendingConnectionSource: null,
-        pendingRemovalEdgeIds: new Set(),
-        pendingRemovalNodeIds: new Set(),
-      },
-      topology,
-      workstationResolver: {
-        resolveWorkstation: (name) =>
-          factoryWithoutStopWords.workstations.find(
-            (workstation) => workstation.name === name,
-          ),
-      },
-    });
-    const workstationNode = projection.nodes.find(
-      (node) => node.id === "workstation:draft",
-    );
-
-    expect(workstationNode?.data.zAxisIncompleteHints).toEqual({
-      accessibleLabel: hintMessage,
-      title: hintMessage,
-    });
-    expect(
-      projection.nodes.find((node) => node.id === "work-state:story:queued")
-        ?.data.zAxisIncompleteHints,
-    ).toBeNull();
   });
 
   it("omits worker-assignment handles on LOGICAL_MOVE workstations", () => {
@@ -489,48 +436,5 @@ describe("factory graph React Flow projection", () => {
 
     expect(anchorIds).toContain("workstation-on-continue-source");
     expect(anchorIds).toContain("workstation-on-rejection-source");
-    expect(
-      projection.nodes.find((node) => node.id === "workstation:draft")?.data
-        .zAxisIncompleteHints,
-    ).toBeNull();
-  });
-
-  it("omits z-axis incomplete hints when connection editing is disabled", () => {
-    const factoryWithoutStopWords = {
-      ...baseFactoryDefinition,
-      workstations: [
-        {
-          ...baseFactoryDefinition.workstations[0],
-          behavior: "STANDARD",
-          stopWords: undefined,
-        },
-      ],
-    } satisfies CanonicalFactoryDefinition;
-    const topology = buildFactoryGraphTopologyFromDefinition(
-      factoryWithoutStopWords,
-    );
-
-    const projection = projectFactoryGraphToReactFlow({
-      editor: {
-        canEditConnections: false,
-        pendingAdditionEdgeIds: new Set(),
-        pendingAdditionNodeIds: new Set(),
-        pendingConnectionSource: null,
-        pendingRemovalEdgeIds: new Set(),
-        pendingRemovalNodeIds: new Set(),
-      },
-      topology,
-      workstationResolver: {
-        resolveWorkstation: (name) =>
-          factoryWithoutStopWords.workstations.find(
-            (workstation) => workstation.name === name,
-          ),
-      },
-    });
-
-    expect(
-      projection.nodes.find((node) => node.id === "workstation:draft")?.data
-        .zAxisIncompleteHints,
-    ).toBeNull();
   });
 });
